requirements.txt
# This file is for pip compatibility.
# For project metadata, see pyproject.toml.
numpy
pyyaml